Move files for build with SPM old versions
